fix: address PR review feedback - fix type organization
- Move IGalleryListState from client hook to shared types where it belongs
- Fix test import that was importing from wrong location
- Remove unused IGalleryImage import from useGallery.ts

Previous review score: 78/100
